pyproject.toml
[tool.poetry]
name = "test_postgres"
version = "0.1.2"
description = "Context for PostgreSQL database testing"
authors = ["Yevhen Shymotiuk <yevhenshymotiuk@gmail.com>"]

repository = "https://github.com/yevhenshymotiuk/test_postgres"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
